+++ config/Migrations/23_CaixaCaixaMatriz.php
@@ -17,7 +17,7 @@ class CaixaCaixaMatriz extends AbstractMigration
          $table = $this->table('caixa_caixa_matriz');
         $table->addColumn('caixa_matriz_id', 'integer')
                 ->addColumn('caixa_id', 'integer')
-                ->addColumn('peso', 'decimal', ['null' => false])
+                ->addColumn('peso', 'float', ['null' => false])
                 ->addForeignKey('caixa_matriz_id', 'caixa_matriz', 'id', ['delete' => 'CASCADE', 'update' => 'CASCADE'])
                 ->addForeignKey('caixa_id', 'caixa', 'id', ['delete' => 'CASCADE', 'update' => 'CASCADE'])
                 ->create();
